requirements.txt
beautifulsoup4
fuzzywuzzy
python-Levenshtein
python-telegram-bot==13.0
Sphinx
requests
py
apiai